rename/move AuditComponent to ListAuditEntriesComponent
