+++ src/app/core/admin/admin-routing.module.ts
@@ -8,7 +8,7 @@ import { AdminCreateEuaComponent, AdminListEuasComponent, AdminUpdateEuaComponen
 import { CacheEntriesComponent } from './cache-entries/cache-entries.module';
 import { AdminComponent } from './admin.component';
 import { AuthGuard } from '../auth/auth.guard';
-import { AuditComponent } from './audit/audit.module';
+import { ListAuditEntriesComponent } from './audit/audit.module';
 
 @NgModule({
 	imports: [
@@ -70,7 +70,7 @@ import { AuditComponent } from './audit/audit.module';
 
 					{
 						path: 'audit',
-						component: AuditComponent
+						component: ListAuditEntriesComponent
 					}
 				]
 			}])
